+++ pmd-java/src/test/java/net/sourceforge/pmd/cli/CLITest.java
@@ -9,6 +9,8 @@ import static org.hamcrest.Matchers.containsString;
 
 import org.junit.Test;
 
+import net.sourceforge.pmd.PMD.StatusCode;
+
 /**
  * @author Romain Pelisse &lt;belaran@gmail.com&gt;
  *
@@ -51,21 +53,21 @@ public class CLITest extends BaseCLITest {
     @Test
     public void exitStatusWithViolations() {
         String[] args = { "-d", SOURCE_FOLDER, "-f", "text", "-R", "category/java/errorprone.xml", };
-        String log = runTest(4, args);
+        String log = runTest(StatusCode.VIOLATIONS_FOUND, args);
         assertThat(log, containsString("Avoid empty if"));
     }
 
     @Test
     public void exitStatusWithViolationsAndWithoutFailOnViolations() {
         String[] args = { "-d", SOURCE_FOLDER, "-f", "text", "-R", "category/java/errorprone.xml", "-failOnViolation", "false", };
-        String log = runTest(0, args);
+        String log = runTest(StatusCode.OK, args);
         assertThat(log, containsString("Avoid empty if"));
     }
 
     @Test
     public void exitStatusWithViolationsAndWithoutFailOnViolationsLongOption() {
         String[] args = { "-d", SOURCE_FOLDER, "-f", "text", "-R", "category/java/errorprone.xml", "--fail-on-violation", "false", };
-        String log = runTest(0, args);
+        String log = runTest(StatusCode.OK, args);
         assertThat(log, containsString("Avoid empty if"));
     }
 
@@ -75,7 +77,7 @@ public class CLITest extends BaseCLITest {
     @Test
     public void testWrongRuleset() {
         String[] args = { "-d", SOURCE_FOLDER, "-f", "text", "-R", "category/java/designn.xml", };
-        String log = runTest(1, args);
+        String log = runTest(StatusCode.ERROR, args);
         assertThat(log, containsString("Can't find resource 'category/java/designn.xml' for rule 'null'."
                                            + "  Make sure the resource is a valid file"));
     }
@@ -86,7 +88,7 @@ public class CLITest extends BaseCLITest {
     @Test
     public void testWrongRulesetWithRulename() {
         String[] args = { "-d", SOURCE_FOLDER, "-f", "text", "-R", "category/java/designn.xml/UseCollectionIsEmpty", };
-        String log = runTest(1, args);
+        String log = runTest(StatusCode.ERROR, args);
         assertThat(log, containsString("Can't find resource 'category/java/designn.xml' for rule "
                                            + "'UseCollectionIsEmpty'."));
     }
@@ -97,7 +99,7 @@ public class CLITest extends BaseCLITest {
     @Test
     public void testWrongRulename() {
         String[] args = { "-d", SOURCE_FOLDER, "-f", "text", "-R", "category/java/design.xml/ThisRuleDoesNotExist", };
-        String log = runTest(1, args);
+        String log = runTest(StatusCode.OK, args);
         assertThat(log, containsString("No rules found. Maybe you misspelled a rule name?"
                                            + " (category/java/design.xml/ThisRuleDoesNotExist)"));
     }