+++ pmd-test/src/main/java/net/sourceforge/pmd/cli/BaseCLITest.java
@@ -25,6 +25,7 @@ import org.junit.BeforeClass;
 
 import net.sourceforge.pmd.PMD;
 import net.sourceforge.pmd.PMD.StatusCode;
+import net.sourceforge.pmd.internal.util.AssertionUtil;
 
 /**
  * @author Romain Pelisse &lt;belaran@gmail.com&gt;
@@ -79,11 +80,19 @@ public abstract class BaseCLITest {
         }
     }
 
+    /**
+     * @deprecated Use {@link #runTest(String...)}, note that
+     *     it returns the log while this returns the name of a file containing the log.
+     */
     @Deprecated
     protected String runTest(String[] args, String testname) {
-        return runTest(args);
+        return runTest(args, testname, 0);
     }
 
+    /**
+     * @deprecated Use {@link #runTest(StatusCode, String...)}, note that
+     *     it returns the log while this returns the name of a file containing the log.
+     */
     @Deprecated
     protected String runTest(String[] args, String testname, int expectedExitCode) {
         String filename = TEST_OUPUT_DIRECTORY + testname + ".txt";
@@ -100,27 +109,49 @@ public abstract class BaseCLITest {
      * Returns the log output.
      */
     protected String runTest(String... args) {
-        return runTest(0, args);
+        return runTest(StatusCode.OK, args);
     }
 
     /**
      * Returns the log output.
+     *
+     * @deprecated Use {@link #runTest(StatusCode, String...)}
      */
+    @Deprecated
     protected String runTest(int expectedExitCode, String... args) {
+        switch (expectedExitCode) {
+        case 0:
+            return runTest(StatusCode.OK, args);
+        case 1:
+            return runTest(StatusCode.ERROR, args);
+        case 4:
+            return runTest(StatusCode.VIOLATIONS_FOUND, args);
+        default:
+            throw AssertionUtil.shouldNotReachHere("unknown status code " + expectedExitCode);
+        }
+    }
+
+    protected String runTest(StatusCode expectedExitCode, String... args) {
         ByteArrayOutputStream console = new ByteArrayOutputStream();
         PrintStream out = new PrintStream(console);
         System.setOut(out);
         System.setErr(out);
         StatusCode statusCode = PMD.runPmd(args);
-        assertEquals(expectedExitCode, statusCode.toInt());
+        assertEquals(expectedExitCode, statusCode);
         return console.toString();
     }
 
+    /**
+     * @deprecated Use {@link #runTest(StatusCode, String...)}
+     */
     @Deprecated
     protected void runPMDWith(String[] args) {
         PMD.main(args);
     }
 
+    /**
+     * @deprecated Use {@link #runTest(StatusCode, String...)} instead of checking the return code manually
+     */
     @Deprecated
     protected void checkStatusCode(int expectedExitCode) {
         int statusCode = getStatusCode();
@@ -129,6 +160,10 @@ public abstract class BaseCLITest {
         }
     }
 
+    /**
+     * @deprecated Use {@link #runTest(StatusCode, String...)} instead
+     * of checking the return code manually
+     */
     @Deprecated
     protected int getStatusCode() {
         return Integer.parseInt(System.getProperty(PMDCommandLineInterface.STATUS_CODE_PROPERTY));